+++ src/DotNetWorker.ApplicationInsights/release_notes.md
@@ -4,3 +4,4 @@
 
 - Updating `Azure.Identity` from 1.12.0 to 1.17.0
 - Updating `Microsoft.ApplicationInsights.PerfCounterCollector` from 2.22.0 to 2.23.0
+- Improved idempotency of service registration calls (#3273)